requirements.txt
google-generativeai>=0.3.0
tweepy>=4.14.0
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
pydantic>=2.5.0
schedule>=1.2.0 